Day 5 and read file boiler

+++ src/day3_2.rs
@@ -2,7 +2,7 @@ use std::collections::HashSet;
 // use std::env;
 use std::fs;
 
-fn main(file_path: &str) -> isize {
+fn main(file_path: &str) -> u16 {
     // let args: Vec<String> = env::args().collect();
     // let file_path = &args[1];
     let contents = fs::read_to_string(file_path).expect("No file found");
